Rewrite build-release-branch.sh as create-release-branch.sh
The build is supposed to be done by GitHub Actions now, so a lot of the
former logic here can go away. All we need to do is create and push the
branch with the proper naming.

At the same time, new logic was needed to build generically. And I
included a bit more in the way of safety checks while I was at it.