[not verified] Autoloader: refactored test suite
Since the autoloader is intended to be used in a generated state it
makes sense that our tests would be against that. This commit
removes all of the dummy plugins in favor of a solution that
supports generating plugins using a factory.

I also had to rename all of the test suite classes to deal with a
deprecation in PHPUnit 9.5.1 as well as improved the separation
between the different types of tests in the package.