[not verified] Clear the autoloader plugin cache when the Jetpack version is changed (#105)
Both the jetpack and jetpack-dev plugin paths can be added to the autoloader's
plugin paths cache if the user switches between stable and development versions.
This will result in undesirable behavior. The autoloader will see both the jetpack
and jetpack-dev plugins as active and will use the jetpack plugin's packages
when jetpack-dev is active. This happens because the autoloader prefers stable
package versions.

Fix this by clearing the autoloader's plugin paths cache when the Jetpack version
is changed.