Autoloader: revised latest autoloader inclusion semantics (#18354)
The `$jetpack_autoloader_including_latest` global variable was used by the v2.7 and v2.8 versions
of the autoloader to help determine whether the current autoloader was included by another autoloader
rather than loaded by its plugin. This ended up not being robust enough and caused different bugs when it encountered a state it was not prepared to handle.

We've now retired this global in favor of more explicitly inferring the inclusion state from the existing globals. This avoids all of the cases where a bad assumption would lead to fatal errors due to missing classes.