Autoloader: reintroduce the $jetpack_autoloader_including_latest global
When dev versions of Jetpack that use the latest version of the autoloader
are active alongside plugins that use older versions of the autoloader,
fatals can occur. The fatals occur when the JETPACK_AUTOLOAD_DEV constant
is not set to true, so the autoloader prefers stable versions over dev
versions.

The fatals occur because the dev version of the autoloader selects the older
version of the autoloader and attempts to load it. The dev version does not
set the $jetpack_autoloader_including_latest global variable. However, the
older version of the autoloader uses that variable to determine whether it
should initialize.

To fix this, set the value of the $jetpack_autoloader_including_latest in
Latest_Autoloader_Guard::should_stop_init.