[not verified] Prevent upgrade nudge from being clickable while hidden
The upgrade nudge banner is set to `opacity: 0` when hidden, which
makes the banner invisible but does not prevent it from being
clicked. This can be confusing because it can prevent clicks on the
underlying block, with no indication of what's happening.

`visibility: hidden` is chosen so that the banner still takes up
space when hidden, so that blocks do not jump around the page
excessively as the banners are hidden/unhidden.