Also run the workflow when the PR is opened or updated
GitHub sometimes keeps reviews after pushing a merge commit to a PR, but
dismisses the status check. We should re-run in this situation to pick
that back up.

And we may as well run when the PR is created too, to get the status
showing as "Awaiting reviews" with our icon instead of the default
"Expected — Waiting for status to be reported".