Upgrade Flow: Jetpack will not initiate the connection flow.
Jetpack will no longer initiate the connection flow. Instead, we only pass the GET parameter `unlinked=1` to the WP.com redirects component, informing that the user is unlinked and needs to be linked.

WP.com will determine if the link requires the user to be connected, i.e. the user goes straight to checkout or to the landing page first.

If connection is necessary, the redirect component will initiate the connection flow.

Otherwise, the connection flow will be skipped, and the user will be redirected to the destination (the landing page).