Beta plugin: set the JETPACK_AUTOLOAD_DEV constant when changing versions
The Jetpack autoloader prefers to load stable versions of packages over development versions.
This behavior can be overridden by setting the JETPACK_AUTOLOAD_DEV constant to true.

When testing development versions of Jetpack using the Beta plugin, we need to make sure that the
JETPACK_AUTOLOAD_DEV constant is set to true so the development versions of packages will be used.

 - Add a new option, 'jetpack_autoload_dev', which is set when a Jetpack version is activated.
   - Set the option to false when the stable or release candidate versions are activated. These
     versions use stable versions of the packages.
   - Set the option to true when development versions are activated.

 - Set the JETPACK_AUTOLOAD_DEV constant using the value of the 'jetpack_autoload_dev' constant when
   the Jetpack Beta plugin is loaded.