codesniffer: Hack around mediawiki-codesniffer bug
Their test for whether a class is a PHPUnit test class is mainly
designed for their own use, and doesn't after all work for us.
Fortunately they have that logic isolated in a trait, which we can copy
and override.

Also fix the one file that was removed from the excludelist thanks to
this bug (https://phabricator.wikimedia.org/T273575).

And fix three other files that were removed from the excludelist thanks
to a different bug that broke detection of `dirname( __FILE__ )` in some
cases (https://phabricator.wikimedia.org/T273574).