Rewrite JS test workflow
Again, this renames the checks, so we'll have to disable the check
requirements for the old names and enable it for the new.

This also drops the Jetpack "Build dashboard & extensions" test
entirely. We're already doing a build for the Build workflow's "Build
all projects" job, so we may as well just mark that as required instead
of running a separate test.