Rewrite coverage test workflow
Surprisingly didn't rename the job for this one.

It's not entirely clear to me how the two different code coverage
websites we upload to are handling the different paths in the different
files. It would be good to check that the changes I made didn't break
either one of them.