CI: Make tests more generic (#18685)
The general idea here is that each project should define composer commands for the different types of tests we support, and the infrastructure should just run them without doing too much that's specific to any one plugin.

Thus, the PHP, JS, and Coverage workflows are rewritten to move all the Jetpack-specific stuff into Jetpack, and as a side effect the packages also get some changes to be more flexible than they had been previously.

I also added some logic to better avoid running unnecessary tests: A change that only touches the Beta plugin, for example, probably doesn't need Jetpack's tests run. But a change that touches the Autoloader package will also need Jetpack tests to ensure that it didn't break Jetpack.

I punted for now on the E2E tests. They're done in their own weirdly complicated way so it'll take more work to figure out which parts can be made generic, which are Jetpack-specific, and which if any we might be able to get rid of entirely.