eslint-changed: Use merge-base to fetch old file versions
We're using `git diff gitBase...HEAD` to get a diff for line number
correction, which uses the merge-base of gitBase and HEAD as the left
side of the diff rather than gitBase itself.

But when we run eslint, we run it against the file versions from gitBase
itself rather than from the merge-base commit.

This can break the line number correction and result in unchanged issues
being reported as changed.