Fixing the broken phpunit tests. I learned the hard way that the test class file name needs to be prefixed with "test-".
Fixing a bug in the get_other_user_data() method related to type checking.
Basically, the $user_data was expected to always contain an object.
When the user functions are mocked, the $user_data variable is set to false.
I just added some checking of the object vars before trying to use them.