Actions: switch from app to action for Stale issue/PR labeling
Instead of relying on an app, let's use actions to mark issues and labels as stale.
I've kept our stale options as close to the original as possible. The only option we're losing (because it is not available in the action) is the exemption for issues and PRs that are part of a project. I think it's an acceptable change.