Parser for keepachangelog.com format
Went back and forth a bit on whether to pull in a markdown parser or
write my own thing. What decided it in the end was that I couldn't find
one that offered parsing to an abstract syntax and converting that AST
back to markdown, plus most didn't support PHP 5.6.

Haven't actually tested this yet. That's for tomorrow.