CLI: Fix CodeQL warning
We tried to match the regex documented at
https://getcomposer.org/doc/04-schema.md#name for allowed package names.
But somehow we lost the `/` in the middle. We really only wanted the
part after the slash: `^[a-z0-9](([_.]?|-{0,2})[a-z0-9]+)*$`

But that wouldn't satisfy CodeQL. We could simplify that to the more
efficient `^[a-z0-9]+(([_.]|-{1,2})[a-z0-9]+)*$`

But that's not what Composer is actually using (or, AFAICT, has ever
used). They just use `^[a-z0-9_.-]+$`, see
https://github.com/composer/composer/blob/03e8cacd/src/Composer/Command/InitCommand.php#L239