Revert ToolbarItem changes for the Whatsapp block
Since that block has a set of tests, the tests need to be able to use ToolbarItem. If I update the version of @wordpress/components so it includes ToolbarItem, the tests pass, but the overall block build fails:

    ERROR in ./node_modules/@wp-g2/create-styles/node_modules/postcss/lib/map-generator.js
    Module not found: Error: Can't resolve 'path' in '.../jetpack/node_modules/@wp-g2/create-styles/node_modules/postcss/lib'

Let's consequently hold off this change until @wordpress/components is properly updated in a separate PR: #16763.