Passing the iframe height and source to WP.com.
To properly handle styling inside the iframe (on the WP.com side), we add two new parameters: `iframe_height` and `iframe_source`, so we knew the size of the iframe, and where exactly in Jetpack it's being loaded.

This commit also makes some minor styling adjustments to the connection screen and the user connection bar.