Build: Build for forked PRs
We currently trigger build on "push", which runs in the context of the
forked repo and so the PR doesn't see it as having run.

So also run on PR, but only for forked PRs.

Also fix it so the Beta Plugin job doesn't try to run for pushes to
forked repos, since it won't work anyway.