Linting: Cleanup test skipping
PR #18173 made the mistaken assumption that booleans in job outputs
would be translated in a way such that `if: output` would work. In fact,
booleans in outputs become string "true" or "false" and `if` interprets
string-"false" as truthy. This was causing tests not to be skipped as
intended.

Also, remove the hack from #17940 and just run the linting job always.
The hack doesn't work for forked PRs.