Tests: Cleanup CodeClimate env logic
For fork PRs, the head ref doesn't exist in "origin". And GH already
provides the needed SHA anyway, so just use that.

Also, it seems that CodeClimate wants "master" rather than
"refs/heads/master" that GITHUB_REF contains, so fix that too.

And while we're at it, make sure the env vars are set for the call to
./cc-test-reporter before-build. Apparently they currently don't care,
but someday they might.