Theme Tools: Updating the logo in Appearance Customizer.
When site logo is replaced using the Customizer, the logo image in the site preview does not change.

Jetpack replaces the `src` attribute of the image with new value, but `srcset` remains the same, so the browser continues to display the older image.

This commit makes Jetpack remove the `srcset` attribute from the site preview logo when new logo gets chosen, so the updated `src` attribute takes effect immediately.