linting: Increase time limits for eslint jobs (#19050)
PR #18504 increased the limit for one job that was consistently hitting
the old 5-minute timeout, but not the other two. Lately I've been
noticing that one of those other two has been barely hitting the timeout
too on occasion, and I see the other is at ~4:30 too, so let's bump them
too.