Actions: remind authors to commit their wpcom diffs with comment.
As an a12n, once you merge a PR in Jetpack, it's nice to follow-up and also commit any matching WordPress.com diff that was created for you.
This commit adds a new task that will run every time a push to master happens, will look for the matching merged PR, and will add a reminder comment if that PR also included a previous comment from matticbot.