Replace dev-monorepo with dev-master everywhere
In #17259 we added some hacks to avoid Composer switching from the
monorepo package to the dev-master available from packagist. Since then,
Composer 2 has been released which has different behavior that will not
have the same problem.

And meanwhile, we'd have to start adding "dev-monorepo" in a lot more
places in order for the move to using branch-alias to work. Instead,
let's switch back to using dev-master everywhere.