cli: Use `composer install` or `composer update` as appropriate
If the lock file isn't checked in to git, we should use
`composer update` instead of using whatever stale lockfile happens to be
in place.

And the same for yarn. But yarn needs to use `install` rather than
`upgrade` when no lockfile exists at all, while composer wants the
opposite.

Also it looks like the `opts` weren't being passed to Listr properly and
so weren't being respected. It seems it picked up the renderer from the
parent Listr, which is fine, but was always running in non-concurrent
mode.