linting: Add check that yarn.lock is up to date
Much as we do for composer.lock, let's also ensure that any checked-in
yarn.lock files are up to date. Yarn seems to make it a little easier
than composer does, as `yarn install` doesn't blindly trust the lock
file and it even has a `--frozen-lockfile` option to directly indicate
when an update is needed.

While doing so, let's merge the "Composer" workflow into "Linting" since
the cleanest place to add the yarn.lock check is in the same job as the
existing composer.lock check.

This renames one of the required workflows, so merging will require
disabling the old name and enabling the new.

Fixes #19027