Jetpack tests: Fix flaky shortcode tests by mocking the http requests
These aren't supposed to be integration tests against the live data from
the remote providers, and are often the cause of spurious test failures
when the upstream provider has an error or something.

Some of the tests already have their own request mocking logic; I added
the new thing in there anyway in case future added tests aren't mocked.

This will probably also make the tests a bit faster, as it won't be
making a bunch of http requests.