Add new Design input task
This task runs every time a PR is labeled. If a Design status label is added, it posts to the Design channel accordingly, and then adds a new label to the PR to show that input was requested. This avoids having the bot post again every time the PR is labeled.