SEO: ignore instance of react/jsx-no-bind errors
From commit 383f4d3, when the token buttons were clicked while the
handleTokenButtonClick() function was wrapped in useCallback, it was
creating state issues with the other inputs. There may be a better
approach to use here, but for now I'll ignore the one instace of using
arrow funcations in props.