Do not clear the GUID when the fetch 404s.
Clearing the GUID will cause the placeholder to render on the frontend as
well, but only after the user saves the post. So if a user deletes a video
file from their Media Library which is used by a block in an old post,
that block displays on the frontend as a broken link. Only if the user
happens to open the post in the editor and save it (without actually fixing
the block) will the placeholder get rendered on the frontend instead.

This commit preserves the GUID instead, so that the behavior on the frontend
will be consistent.