Use decodeEntities for SEO input value and preview
Since we are sanitizing with `esc_html`, some input gets encoded to HTML
entities which aren't so nice to look at when editing in the inputs or
in the live previews. We can decode the string with HTML entities before
handing it off to React which will escape by default.