actions: Correctly supply git auth token (#19442)
Apparently GitHub sets up a generic auth token by default, and what we
were doing wasn't actually valid so it was falling back to that. This
should be the correct way to set it up.

This mostly doesn't matter, since either token is allowing access in
our current use cases. But it does prevent cancel-dups.yml from being
triggered from renovate-helper.sh and might break push-to-mirrors if
someone were to try using it to push to a private repo.