Search: use the posts_per_page value from the query
Currently, the filter__posts_pre_query method uses the posts_per_page option value to determine the number of
posts to return. Instead, it should use the posts_per_page value from the input $query to determine the number of posts to
return. This will allow sites to use the `pre_get_posts` filter to adjust the number of returned posts.