Photon: Cast dimensions to int when comparing (#19538)
* Photon: Cast dimensions to int when comparing

WordPress is not particularly type-safe. In 3908038-zen I ran into a
situation where a site had the width and height metadata in 'sizes'
stored as strings rather than integers. WP was passing those strings
through in the srcset data, and since #16588 made a bunch of comparisons
strict Photon wasn't matching them anymore.

So add some casts to fix that.