tests: Stop setting GITHUB_TOKEN when running tests
It's not even passing a token, it's passing the literal string
"secrets.API_TOKEN_GITHUB" due to incorrect syntax.

If it were passing a token, whatever tests were trying to use it would
probably fail on forked PRs. If we do wind up needing a token for tests,
we should try to use GH's own `secrets.GITHUB_TOKEN` (which will always
exist, but with read-only access for forked PRs) instead of matticbot's.