[not verified] # This is a combination of 3 commits.
I don't like doing this, but there's no other way to enqueue/inject the styles we need for the podcast player into the site editor iframe.
So, using JS and the DOM, we find the CSS we need from the head of the parent window and move it into the site editor.
YAY.

Changelog

Adding null checks to the elements

we should ensure that the parent-level dom elements are there before we use them