[not verified] Admin Menu: Fix invalid menu item ID attributes
Some menu items are pointing to full URLs. It turns out these have a
number of invalid characters that aren't getting stripped out by the
core WordPress functions. This meant that invalid ID attributes were
being set and that in turn broke the SVG styling introduced in #19364.

This PR fixes the ID, allowing the icon to be properly displayed.