[not verified] Connect: add modal when dismissing connection banner
Internal reference: p1HpG7-bPy-p2

This modifies the behaviour of the connection banner's dismiss button.
Instead of just dismissing and hiding the banner, we would like to deactivate the plugin entirely. If you do not intend on connecting your site to WordPress.com, there is no point in keeping the Jetpack plugin active.

This work is partially reusing elements of the deactivation modal introduced in #15858, in order to avoid duplicating code.

This first commit is still a work in progress:
- I need to fix linting errors in the js
- Right now this only works from the plugins page. That's an issue since that banner also appears in the main dashboard. I'll explore using the `/wp/v2/plugins/jetpack/jetpack { status: "inactive" }` in a future commit to have a solution that doesn't rely on fetching the deactivation link from the page.
- The modal's style is not perfect and still needs to be adjusted.