[not verified] Move asset enqueueing to render_block
For FSE sites, it is no longer reliable to assume that assets should
be enqueued when the Related Posts feature is enabled, because the
actual block is not necessarily present on the page. This commit
moves the enqueueing of assets to the `render_block` method instead,
so that they are only enqueued when necessary. This also allows us
to remove the null checks in the scripts.

For Related Posts that are inserted via the hook on non-FSE sites,
this would break the client-rendered posts for non-AMP requests, so
we also need to enqueue assets within that branch.