build: Fix package ref updating (#19670)
We used `composer update --root-reqs` to try to update references to
monorepo packages to point to the built versions instead of the monorepo
versions.

But that fails to update any monorepo packages that are depended on
indirectly, like how plugins/jetpack depends on packages/sync which
depends on packages/password-checker but does not itself depend on
packages/password-checker.

This probably doesn't matter much, except for the case where we're
bundling the packages in the plugin. In that case we might get files
that aren't supposed to be included, or omit files that are.

So instead, explicitly update just the packages built from the monorepo.