JITM: set the JITM and cache filter defaults to true (#19570)
* JITM: set the JITM and cache filter defaults to true

The 'jetpack_just_in_time_msgs' filter is used to enable and disable the display of the JITMs. Let's
set the default value to true. Currently, the plugin has to set the filter to true in order to display JITMs
If a plugin has configured the JITM message package, it's safe to assume that the JITMs should be displayed.

Also, the 'jetpack_ust_in_time_msg_cache' filter is used to enable and disable the JITM cache. Let's
set the default value to true.

Note that the Jetpack, currently the only consumer of the JITM package, currently sets both of these filters to true.