[not verified] Remove unauthorized check of parent window location
Looks like this check was originally used by the upgrade nudges to identify
when the block was in Calypso and prevent redirection. After adding in the
checkout modal it appears to be unnecessary (and upgrade specific).

This check was also causing CORS issues for Stripe nudges in the site
editor, due to the cross-origin iframe. Removing it resolves this bug.