changelogger: Find base dir by looking for composer.json (#19721)
* changelogger: Find base dir by looking for composer.json

Like composer itself does, look for composer.json in parent directories
if the current directory does not have one. If composer.json is found in
the parent, ask whether it should be used (and exit if the user answers
"no").

This also cleans up Config to no longer need an OutputInterface
injected, which I've never really liked but hadn't seen a way around it
before. Now the way around it is obvious.

* Everything depends on changelogger, so everything needs a dep update

* Update root composer.lock too

* Old versions of symfony/console lack ApplicationTester::setInputs()