JITM: move the last sync callback hook to the JITM package (#19730)
* JITM: move the last sync callback hook to the JITM package

The jetpack_last_plugin_sync transient value is used to determine if the JITM cache should be used.
This transient is updated when the jetpack_sync_before_send_updated_option filter fires. Move the
jetpack_sync_before_send_updated_option hook and the jetpack_track_last_sync_callback method from
the Jetpack class to the JITM package. This will allow the JITM package to update the
jetpack_last_plugin_sync transient value when the Jetpack plugin is not active.