tools: Check for jq before using it in check-copied-files.sh (#19763)
When that script is run from the git precommit hook, we shouldn't fail
the commit just because jq is missing as it's not exactly a standard
tool. But print a warning about it.

When it's run in CI, however, we do want to fail if jq is somehow
missing. Not that several other CI checks won't also fail in that
case, but the nicer error message is probably good.