changelogger: Handle reused branches in non-interactive mode
If changelogger is being used non-interactively, it may be that the
branch name is being reused and so the automatic filename would result
in an error.

Add a command line option that indicates that this is expected and a
numeric suffix should automatically be added to avoid an error.

We don't do this in interactive mode. When a human is using the tool,
they should be informed and allowed to choose an alternative file name.