tools: Avoid `composer update --root-reqs` in more places
In tools/check-composer-deps.sh and the CI lockfile check, we were using
`composer update --root-reqs` to try to update monorepo packages in
composer.lock.

But that fails to update monorepo packages that are depended on
indirectly. What we need to do is pass the list of all monorepo packages
(and the root reqs) to `composer update` so it can update those and only
those.

Since the code to properly determine that list is somewhat complicated,
let's wrap it in a script.