[not verified] AMP: Standard mode disables the WPCOM toolbar in wp-admin
Whilst developing #19525 we discovered that the WPCOM toolbar module was
disabled in the admin when the AMP template mode was set to standard.
This is because `amp_is_request` always returns `true`. This change
fixes this by adding an additional check for AMP being available, which
it is not in admin, CLI requests etc.