Admin Page: display Sharing as new default tab for non-admins
This commit changes the behaviour of the Settings screen for connected non-admins.

Non-admins currently need access to the Settings screen for 2 reasons:
- Publicize (Sharing tab)
- Post By Email (Writing tab)

Until now, we always defaulted to the Writing tab for connected non-admins. This has 2 problems:
- We did not check if the Post By Email module was actually enabled, so when the module was disabled editors loading the settings page would only see a Writing tab and its title, but no content.
- Publicize is more popular than Post By Email today, so it's more likely that editors will be interested in the Publicize settings first, even if both PBE and Publicize are enabled.

This commit changes that default behaviour, and introduces checks for the active modules to avoid displaying tabs when we do not need to.

Previous work on this: #13541