ci: Use new "concurrency" feature instead of potiuk/cancel-workflow-runs (#19855)
GH has recently added the ability to declare concurrency groups for
workflows or jobs, which allows for cancelling pending (or running)
instances of the workflow/job when a new instance is started.

We can easily enable this for the CodeQL, E2E, Linting, PHPCompat,
Renovate helper, and Tests workflows, setting the "group" for each based
on the event and ref.

We can also enable it for Build, but it's a little more tricky since we
do not want to cancel the runs on push to master or release branches,
only to PRs. So let's reconfigure it so all PRs run for the
"pull_request" event and only the pushes to master and release branches
run for the "push" event. Then for push we can give each run a unique
group by including the run_id.

Then we can get rid of the Cancelling workflow, which should save us a
good bit on our API call limit.

I didn't bother enabling it for PR is up-to-date, Renovate
cleanup, Check required reviews, or Mark stale issues. These are all
unlikely to have concurrent runs, either through being scheduled
or being fast enough.

I also didn't bother enabling it for Gardening, as we generally do want
that to run when duplicated (plus it's not slow either).