Replace "yarn" with "pnpm" in docs
As far as what's listed in the docs, it turns out pnpm is mostly a
drop-in replacement. The main differences are:

* You have to use `pnpm` for scripts versus `pnpx` for bins, while
  `yarn` handles both.
* When running a script via `pnpm`, you need a `--` before any options
  that are supposed to be passed to the script. `yarn` does not require
  this.

I also saw a lot of places where the docs are outdated, e.g.
"Development workflow" in docs/development-environment.md seems to have
never been updated for the changed monorepo layout. Fixing that seems
out of scope for this PR.