Beta: Remove the jetpack_autoload_dev option and the JETPACK_AUTOLOAD_DEV constant
The package versions in the autoloader classmap file in the Beta plugins's development versions of Jetpack
are now in a comparable format, such as 1.1.9999999.9999999-dev. Therefore, it's no longer necessary to set
the JETPACK_AUTOLOAD_DEV constant when a development version of Jetpack is selected in the Beta plugin.

The jetpack_autoload_dev option was used to determine whether the JETPACK_AUTOLOAD_DEV constant should
be set, so remove it.