Respect auto_update_plugins option  https://github.com/Automattic/wpcomsh/pull/697
Here we stop forcing every WoA site to auto-update all plugins.
But we want all plugins to auto-update by default, so we keep track of
the most recent plugin list we've seen and use that to detect and add
new plugins to the auto update list. Then, if users want to disable
auto-update for those plugins, they can, and we will respect that.