build: Produce plugins artifact holding zips for all plugins (#20137)
This is the first step towards allowing the Jetpack Beta Tester plugin
to handle other plugins besides just Jetpack.

The build will still also produce the jetpack-dev artifact, so as to
continue working with the existing Beta Download website. That will be
removed in a future PR, after the Beta Download website has been updated
to download the new artifact instead.

This also moves Jetpack's to-test.md from the monorepo root into Jetpack,
so that other plugins can add their own to-test.md in the future.