Janitorial: clean up deprecated code
Following #16433, let's remove all code that was deprecated more than 6 months ago.
This commit also standardizes some of the version numbers so they can start getting the proper notices, and we can then remove the deprecations in the future.

Note that "6 months" here refers to 6 months as of the last time this PR was updated, in December 2020.