Replace most uses of Utils::is_network_active()
The menu option wasn't being added if Jetpack (not Beta) was
network-activated, I guess for access control. In a multiple-plugin
world, we'll need to do that access control inside the admin page based
on the plugin selected.

Similarly, Utils::admin_url() was generating links to regular Admin or
Network Admin based on whether Jetpack (not Beta) was network-activated.
That too will need to depend on if it's generating a link to the
"manage" page for a network-activated plugin or not.

Several hooks needed updating for a multiple-plugin world where some
might be network-activated and some might not be.

To avoid potential network hits on every page load from those hooks if
transients are being particularly transient, cache a map of known plugin
files in the options.

There's still one use left in Utils::replace_active_plugin(); that needs
refactoring anyway, so I'll get rid of the last use when I do that.