pre-commit hook: Only lint changed PHP files
PR #19334 intended to add linting of changed PHP files in the pre-commit
hook, but the actual implementation linted all PHP files if any PHP file
is changed.

This fixes that, by enhancing our wrapper script to be able to be told
when files are being listed on the command line so it can avoid
processing all files.