File Uploads: Prevent unsupported file uploads on lower plans https://github.com/Automattic/wpcomsh/pull/706
* File Uploads: Prevent unsupported file uploads on lower plans

Currently Atomic sites are treated like Jetpack sites, so all files are
trusted for upload to the media library. This means that if the site
doesn't have a plan then files are still uploaded to the site that
wouldn't be allowed on the equivalent Simple site.

This change checks for an unsupported plan, and then restricts the
allowed mimetypes to the same list as WPCOM.

The problem with this approach is that the list is hardcoded so would
need to be kept in sync with WPCOM, and this will also restrict uploads
on Premium and Personal plan sites. The latter is not currently an
issue, as we don't support these plans on Atomic, but ideally we would
request the allowed mime types from the WPCOM API. This should be
handled in a follow up change.

* Change filter priority, to ensure we'll run first.