[not verified] Enqueue assets as early as possible
If assets are enequeued too late, users will not be able to hook
in to dequeue them and replace them with custom scripts.

Since assets need to be enqueued early, add additional checks to
make sure they are not enqueued unnecessarily (when the feature will
not be inserted, because the block already exists in the post or
the site is running an FSE theme).