Beta plugin: Support managing multiple plugins (#20168)
* Cleanup constants
* Create Plugin class
* Clean up AutoupdateSelf clas
- Declare a field that was being used.
- Fix up doc blocks.
- Remove some unnecessary two-line functions, which lets us get rid of
  another field.
- Remove some unneeded `empty()` calls.
- Simplify a weirdly overcomplicated conditional.
- Privatize methods that aren't used externally.
- Rename a misspelled method, `has_never_version` → `has_newer_version`.
- Fix semantics in `plugins_api` filter method `get_plugin_info()`.
- Document why `upgrader_source_selection` is being filtered at all.
* Rewrite CliCommand
The `activate` command won't work yet, because I haven't written the
method to actually do the installation. Other than that it should work.
* Improve transient deletion on uninstall

If we're going to try to delete transients, let's try to get them all.
So rename the ones from the self-autoupdater to use a common prefix with
the rest of the extension, then select the list of transients from the
database.

* Fix for symlinked installation

* Replace most uses of Utils::is_network_active()

The menu option wasn't being added if Jetpack (not Beta) was
network-activated, I guess for access control. In a multiple-plugin
world, we'll need to do that access control inside the admin page based
on the plugin selected.

Similarly, Utils::admin_url() was generating links to regular Admin or
Network Admin based on whether Jetpack (not Beta) was network-activated.
That too will need to depend on if it's generating a link to the
"manage" page for a network-activated plugin or not.

Several hooks needed updating for a multiple-plugin world where some
might be network-activated and some might not be.

To avoid potential network hits on every page load from those hooks if
transients are being particularly transient, cache a map of known plugin
files in the options.

There's still one use left in Utils::replace_active_plugin(); that needs
refactoring anyway, so I'll get rid of the last use when I do that.

* No need to manually handle error_reporting in custom_error_handler

Just return false, the standard handler does the right thing (since PHP
4.3.10!).

* Replace Utils::install_and_activate() (mostly)

Creates a PluginInstaller class to hold that logic instead.

One call remains, from the UI.

This also gets rid of Utils::update_plugin(), also in the UI.

And we clean up Hooks::deactivate() to handle multiple plugins.

* Bugfixes, and add rc and releases to the CLI

* Make Utils::admin_url() take an array

And link to Network Admin if applicable.

* Cleanup hooks (and a few other things)

Changes of note:

* Got rid of the JETPACK_GREEN constant, because we need different
  greens in different places.
   * Logo changed to Green 40, per https://git.io/JcWou
   * Admin bar highlight color changed to Green 50, per pcdRpT-if-p2
* The admin bar menu gets an extra level.
* The autoupdater email loses the attempt at "what changed?" and testing
  instructions. It would be too much if multiple plugins get updated.
  Instead, each plugin listed has a link to the PR or mirror repo branch.
* Pulling in composer/semver to compare versions, since PHP's
  `version_compare()` doesn't really work right.

* Rewrite most of Admin

I decided to more heavily use "template" files rather than embedding
`<?php ?>` in the class file.

I still need to do something about Admin::to_test_content() and
Utils::what_changed(), and go through to remove now-unused methods.

* Fix construct that only works in PHP 7.0+

* Fix Admin::to_test_content()

And merge Utils::what_changed() into it.

Also, instead of using Jetpack's markdown library when it happens to be
available, just pull in one that explicitly handles GitHub-Flavored
Markdown.

* Remove another Jetpack-specific call

* Delete unused stuff

* Add direct-access guard to all template files

* Merge PluginInstaller into Plugin

It didn't turn out to be as big as I thought.

* change entry

* Sort deps in composer.json

* Minor cleanup

* Allow for unpublished plugins

* Handle if .jpbeta.json goes missing

For example, if the -dev dir was left from an older version of Jetpack
Beta.

* Only run in Network Admin if network-activated

Mostly to prevent the possibility of accidentally managing a
network-enabled plugin when our hook for preventing duplicate versions
can't run in time.

* Add missing require